Error found, check it.
In pop() function, you must return an integer in the if statement, you
can see always returns 0 now. I will try to get a solution to get the
display without getting that 0, check the code.

Also I have changed the top=NULL for this->empty(), is the same so you
can change it if you want.